refactor components for staged loading
