+++ src/index.ts
@@ -1,8 +1,8 @@
-export { PopUpModal, MessageType } from "./components/common/PopUpModal";
-export { PoweredByDispatch } from "./components/forums/PoweredByDispatch";
-export { ForumContent } from "./components/forums/forum/ForumContent";
-export { DispatchProvider, useForum } from "./contexts/DispatchProvider";
-export type { DispatchAppProps } from "./contexts/DispatchProvider";
-export { MainForum } from "./utils/postbox/postboxWrapper";
-export { TopicView } from "./views/TopicView";
-export { ForumView } from "./views/ForumView";
+export { PopUpModal, MessageType } from './components/common/PopUpModal';
+export { PoweredByDispatch } from './components/forums/PoweredByDispatch';
+export { ForumContent } from './components/forums/forum/ForumContent';
+export { DispatchProvider, useForum } from './contexts/DispatchProvider';
+export type { DispatchAppProps } from './contexts/DispatchProvider';
+export { MainForum } from './utils/postbox/postboxWrapper';
+export { TopicView } from './views/TopicView';
+export { ForumView } from './views/ForumView';
